+++ src/main/java/lk/ijse/offbeatceylon/repo/AddPlaceRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AddPlaceRepo extends JpaRepository<AddPlaces,Long> {
+public interface AddPlaceRepo extends JpaRepository<AddPlaces,Integer> {
     boolean existsByPlaceName(String placeName);
     AddPlaces findByPlaceName(String placeName);
     AddPlaces findByPlaceId(int placeId);
